Extract listings from seller.js into own file
